+++ bioconda_utils/cli.py
@@ -7,10 +7,11 @@ Bioconda Utils Command Line Interface
 # ".../importlib/_bootstrap.py:219: RuntimeWarning: numpy.dtype size \
 # changed, may indicate binary incompatibility. Expected 96, got 88"
 import warnings
+from bioconda_utils import bulk
 
 from bioconda_utils.artifacts import upload_pr_artifacts
 from bioconda_utils.skiplist import Skiplist
-from bioconda_utils.build_failure import BuildFailureRecord
+from bioconda_utils.build_failure import BuildFailureRecord, collect_build_failure_dataframe
 warnings.filterwarnings("ignore", message="numpy.dtype size changed")
 
 import sys
@@ -184,7 +185,7 @@ def get_recipes(config, recipe_folder, packages, git_range, include_blacklisted=
         all_len = len(recipes)
         recipes = [recipe for recipe in recipes if not skiplist.is_skiplisted(recipe)]
         if all_len > len(recipes):
-            logger.info("Ignoring {all_len - len(recipes)} skiplisted recipes.")
+            logger.info(f"Ignoring {all_len - len(recipes)} skiplisted recipes.")
 
     logger.info("Processing %s recipes%s.", len(recipes),
                 utils.ellipsize_recipes(recipes, recipe_folder))
@@ -437,7 +438,7 @@ def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
           build_image=False, keep_image=False, lint=False, lint_exclude=None,
           check_channels=None, n_workers=1, worker_offset=0, keep_old_work=False,
           mulled_conda_image=pkg_test.MULLED_CONDA_IMAGE,
-          docker_base_image='quay.io/bioconda/bioconda-utils-build-env-cos7:{}'.format(VERSION.replace('+', '_')),
+          docker_base_image=None,
           record_build_failures=False,
           skiplist_leafs=False):
     cfg = utils.load_config(config)
@@ -458,6 +459,14 @@ def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
             use_host_conda_bld = True
         else:
             use_host_conda_bld = False
+
+        if not utils.is_stable_version(VERSION):
+            image_tag = utils.extract_stable_version(VERSION)
+            logger.warning(f"Using tag {image_tag} for docker image, since there is no image for a not yet release version ({VERSION}).")
+        else:
+            image_tag = VERSION
+        docker_base_image = f"quay.io/bioconda/bioconda-utils-build-env-cos7:{image_tag}"
+        logger.info(f"Using docker image {docker_base_image} for building.")
 
         docker_builder = docker_utils.RecipeBuilder(
             build_script_template=build_script_template,
@@ -1008,20 +1017,86 @@ def autobump(recipe_folder, config, packages='*', exclude=None, cache=None,
         git_handler.close()
 
 
-@arg('recipe', help='Path to recipe that shall be skiplisted')
-@arg('reason', help='Reason for skiplisting')
-@arg('--platforms', help='Platforms to skiplist for', nargs='+', type=str, default=['linux-64', 'osx-64'])
-def skiplist_recipe(recipe, reason, platforms=None):
+@arg('recipes', nargs="+", type=str, help='Paths to recipes that shall be skiplisted')
+@arg('--skiplist', action="store_true", help='Skiplist recipes.')
+@arg('--reason', help='Reason for skiplisting. If omitted, will fail if there is no existing build failure record with a log entry.')
+@arg('--category',
+     help='Category of build failure. If omitted, will fail if there is no existing build failure record with a log entry.',
+     choices=["compiler error", "conda/mamba bug", "test failure", "dependency issue", "checksum mismatch", "source download error"]
+)
+@arg('--platforms', help='Platforms to annotate', nargs='+', type=str, default=['linux-64', 'osx-64'])
+@arg('--existing-only', help="Only annotate already existing build failure records. The platform setting is ignored in this case.", action="store_true")
+def annotate_build_failures(recipes, skiplist=False, reason=None, category=None, platforms=None, existing_only=False):
     valid_platform_names = set(conda.base.constants.PLATFORM_DIRECTORIES)
-    for platform in platforms:
-        if platform not in valid_platform_names:
-            logger.error(f"Invalid platform {platform}, choose from: {', '.join(valid_platform_names)}")
-            continue
-        failure_record = BuildFailureRecord(recipe, platform=platform)
-        failure_record.set_commit_sha_to_current_recipe()
-        failure_record.reason = reason
-        failure_record.skiplist = True
-        failure_record.write()
+    for recipe in recipes:
+        if existing_only:
+            platforms = [
+                platform
+                for platform in conda.base.constants.PLATFORM_DIRECTORIES 
+                if BuildFailureRecord(recipe, platform=platform).exists()
+            ]
+        for platform in platforms:
+            if platform not in valid_platform_names:
+                logger.error(f"Invalid platform {platform}, choose from: {', '.join(valid_platform_names)}")
+                continue
+            failure_record = BuildFailureRecord(recipe, platform=platform)
+
+            if not reason and failure_record.exists():
+                if not failure_record.log:
+                    logger.error(
+                        f"Recipe {recipe} has a build failure record ({failure_record.path}), "
+                        "but no log entry. Please add a log entry or specify a reason."
+                    )
+                    continue
+                if failure_record.recipe_sha != failure_record.get_recipe_sha():
+                    logger.error(
+                        f"Recipe {recipe} has a build failure record ({failure_record.path}), "
+                        "but the recipe has changed since recording the build log. "
+                        "Please specify a reason for skipping or rebuild for updating the log."
+                    )
+                    continue
+
+            failure_record.fill(reason=reason, category=category, skiplist=skiplist)
+            failure_record.write()
+
+
+# TODO add subcommand to list recipes with build failure records descendingly sorted by downloads
+# in case of version subdirs, only list if the latest version also has the build failure record.
+# list how many recipes depend on this and sort by it primarily if inner
+@recipe_folder_and_config()
+@arg('--channel', help="Channel with packages to check", default="bioconda")
+@arg('--output-format', help="Output format", choices=['txt', 'markdown'], default="txt")
+@arg('--link-prefix', help="Prefix for links to build failures", default='')
+def list_build_failures(recipe_folder, config, channel=None, output_format=None, link_prefix=None):
+    """List recipes with build failure records"""
+
+    df = collect_build_failure_dataframe(
+        recipe_folder,
+        config,
+        channel,
+        link_fmt=output_format,
+        link_prefix=link_prefix,
+    )
+    if output_format == "markdown":
+        fmt_writer = pandas.DataFrame.to_markdown
+    elif output_format == "txt":
+        fmt_writer = pandas.DataFrame.to_string
+    else:
+        logger.error("Invalid output format, must be txt or markdown.")
+        exit(1)
+    fmt_writer(df, sys.stdout, index=False)
+
+
+@arg(
+    'message',
+     help="The commit message. Will be prepended with [ci skip] to avoid that commits accidentally trigger a rerun while bulk is already running"
+)
+def bulk_commit(message):
+    bulk.commit(message)
+
+
+def bulk_trigger_ci():
+    bulk.trigger_ci()
 
 
 def main():
@@ -1031,5 +1106,6 @@ def main():
     argh.dispatch_commands([
         build, dag, dependent, do_lint, duplicates, update_pinning,
         bioconductor_skeleton, clean_cran_skeleton, autobump,
-        handle_merged_pr, skiplist_recipe
+        handle_merged_pr, annotate_build_failures, list_build_failures,
+        bulk_commit, bulk_trigger_ci
     ])